moved testcases to the correct places
