+++ sootup.java.sourcecode/src/test/java/sootup/java/sourcecode/inputlocation/JavaSourcePathNamespaceTest.java
@@ -6,20 +6,26 @@ import static org.junit.Assert.assertTrue;
 
 import categories.Java8Test;
 import java.util.Collection;
+import java.util.HashSet;
 import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
+import junit.framework.TestCase;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import sootup.core.frontend.AbstractClassSource;
 import sootup.core.frontend.SootClassSource;
 import sootup.core.inputlocation.AnalysisInputLocation;
+import sootup.core.model.SootClass;
+import sootup.core.model.SourceType;
 import sootup.core.signatures.PackageName;
 import sootup.core.types.ClassType;
 import sootup.core.util.ImmutableUtils;
 import sootup.java.core.JavaIdentifierFactory;
 import sootup.java.core.JavaProject;
 import sootup.java.core.JavaSootClass;
+import sootup.java.core.JavaSootClassSource;
 import sootup.java.core.language.JavaLanguage;
 import sootup.java.core.types.JavaClassType;
 import sootup.java.core.views.JavaView;
@@ -82,5 +88,31 @@ public class JavaSourcePathNamespaceTest {
     // We don't check for a specific number here to avoid breaking tests
     // Whenever we add a test class.
     assertTrue(classSources.size() > 1);
+  }
+
+  /**
+   * Test for JavaSourcePathAnalysisInputLocation. Specifying all input source files with source
+   * type as Library. Expected - All input classes are of source type Library.
+   */
+  @Test
+  public void testInputSourcePathLibraryMode() {
+
+    String classPath = "../shared-test-resources/java-miniapps/src/";
+    JavaProject javaProject =
+        JavaProject.builder(new JavaLanguage(8))
+            .addInputLocation(
+                new JavaSourcePathAnalysisInputLocation(SourceType.Library, classPath))
+            .build();
+    JavaView view = javaProject.createOnDemandView();
+
+    Set<SootClass<JavaSootClassSource>> classes =
+        new HashSet<>(); // Set to track the classes to check
+    for (SootClass<JavaSootClassSource> aClass : view.getClasses()) {
+      if (!aClass.isLibraryClass()) {
+        classes.add(aClass);
+      }
+    }
+
+    TestCase.assertEquals("User Defined class found, expected none", 0, classes.size());
   }
 }
